+++ Formula/a2acli.rb
@@ -4,19 +4,19 @@
 class A2acli < Formula
   desc "Standalone A2A CLI client"
   homepage "https://github.com/a2aproject/a2a-rs"
-  version "0.1.8"
+  version "0.1.10"
   license "Apache-2.0"
   depends_on :macos
 
   on_macos do
     on_arm do
       url "https://github.com/a2aproject/a2a-rs/releases/download/a2a-cli-v#{version}/a2acli-v#{version}-aarch64-apple-darwin.tar.gz"
-      sha256 "05e9d0c1be816c7bc2c720dda62e507e5875f627727dc0d50c68fbda6364e98f"
+      sha256 "2e13804755f8f7802a3b39d3c98236ef42bada633f90d8cc66b419bbac8d4933"
     end
 
     on_intel do
       url "https://github.com/a2aproject/a2a-rs/releases/download/a2a-cli-v#{version}/a2acli-v#{version}-x86_64-apple-darwin.tar.gz"
-      sha256 "fd36f14fba844e08f6ac4545747714cf7756f1eb331c849368a7cfb7fe657fa4"
+      sha256 "8f54844085aa69c59eff2fd2725df0ef9510525c833c829aec61581f13594806"
     end
   end